+++ scripts/selection.js
@@ -7,6 +7,7 @@ let logToConsole = false; // Debug
 const os = getOS();
 const notifySearchEngineNotFound = browser.i18n.getMessage('notifySearchEngineNotFound');
 const mycroftUrl = 'https://mycroftproject.com/installos.php/';
+//const ytDownloadUrl = 'https://www.youtubePP.com/watch?v=';
 const base64ContextSearchIcon =
     'iVBORw0KGgoAAAANSUhEUgAAACAAAAAgCAYAAABzenr0AAAG2ElEQVRYhe2Wa1CTVxrH31o/7ezM7kxndndmv6wjs4aEJCCiOx20sOPYdms7uhBaUbou5Y4JBIGogFxiR7BeqmWgSiARCAlvyA2oEMAABbkZVC6CBAkGMCGBo+jY2W5H/feDwhgToLS7s1/2mXm+vc/5/97/c55zDkX9P9YYQcna3/rwtbsCUusEvIKWM9vS9GIfgZbPOlTzrr+I/s1/S3edpL7/7Mmqb83Z5e3PDL1jsDucIITg3swsdmVqwBXqwUnSPWMn65pZfHUoj0e/+R9R5on17wmLWqzZsnbsSKOxI10No8kMQggIIbg1NgWOgAZXqH+ZOnAFNP4qUt1hRkm3/wJprKtsvlXXdsP8PPtyO1KKW3Cp3gR2XAU6BybQNzyJY2XtCE6n8XexHtxkHbhCHfyTlBgen8bktB1XukeeH71klFAU1q1NGnijsWdkoMJwE4GpKohKjIg8fQU+8XJwkjQ4UdmJwDQ1uEIdAoQ1CExXg82nwU6QY3h8GoqWAXQPWWCdmcWUzYHG3tHhNUFovh1uIITgaGkbdmVoMDFlh3NuHrsytC96Lah5xXI9OAI1QsS14Il1SLxgQEpxC8Ym7y+1iRACTftQ008SlzbcPDg3P79UuLiQc24e+YoucARqF/FFoD05Wkjq+3HH4iq8mHPz85A1XP9sVev7RyefvF58Y9SKkDwdgtNpcJI07gDJWuw8qoLDOedRfDFvjt77bsVWyA03Ml8vMprMCExVgStQuVm/mOxD1bBM2yFvHkCQSI2LtSb0DU/CMm13g6gw3MxeFqCt3zzz6sdD41Pg8mmPoi4AfBqn6W6klxiRXtKKwMNK7DyiQvjJOlQbB10A2vvNNo/iF02mX9lmnc8JIbA7nDDfsyH4iObFXK8CsPOoBuNW25JIU98YdB23Uay/jsaeOy4AdocTNN36azeAauNwiN3hxLGydgSmqhBRUO+x326ZpML125PL9r170IJRywwIITgubUdjzx2UNfQfcANQto0UXL89CU6iAjvSVODwVeAka1cFiD1vWHHjTdkcOKXsAiEEIxMzOFHZiYDEqjA3gKyK3mOWaTuumsxIu2R8ueFWt/9zeeeKAIQQlNT3o2fIggmrDXvyasHm0wfdAHxT9LwgkQb5imuYmLLDT1CN0M/r8G6GFuxD1cu6kVvesSqAZdoORcsA9ufXgSvUgRUr/9QNgCVQBy+e53vFtRBXdMA268SsYw53rTb4CapfnveuAFuEKnQOTIAQgvt2Jx5MGrBgEuHRtQgsdEfh4dA5PJgdByEEiYXN4Cbr4P2Z7AM3gD8l0H9g81VLC4fn17v8xYB5Cu+I1B7bEpimRvSZOnxTcQDzjdsw0RyHvvoM3GoUwXl1Lx5f3Y67tzTwFdBg81XYFFGyweMoboorv/viXte4ze/i1ZtU3AKuQOUGoSiLwpguCB9FJyP3TDEKCiUoKJQg/6tLGGzKxAPDNoRlfw1mXKXVozhFURQzsvQ0R1ADNl+FniHLsj39pmsUnFfc2nu8BI8MAQhJTIZ3aCaS8i4sARQUSpBy4itoSj+GsSoE3tHSL5cF8PrHxY2MWNlTrlALkaR1WYDz6l6XTXmmMA2mmt3wDs0Ak5eF8MMFLgBC8QXsEx7GQlMAorJO+i8LQFEU5R0tLfVJUICbVIOa1iGPALtzal3svyyJg748Asyw4/DmZSIu65wLwLFTRXg74jAeN23BfJ0/Y0WAP35a+BYzWnaffagaXIEKXYOurZibm0fwEdeRPF8kRBe9B0xeFrx5mYjNPLsknnv2a3BCRdgTk/DkcdMWzGgYb60IQFEU9eeY0kBmZNn3rPhK1HaOuLwN9opr3Y7oA3mFWGgKwHsxR8AMO47348Qu9jM+TH7aIQtqfWTwN60qvhiMf5btZkRJ/3VK3rYEcKV71OODhCvUo1n+MfpV7+Ptgxnw/SQTBYUSiL+8iG370p9+kfmh4WHj5udmyebYnwxAURTlFVX0l6qmvieEEAyarQjN1S57PG9Pr0Yf/RGsde/g7Lk4FJWeRmpuEhnXbm9baNz8rCPPFzXhvs6qfUzWmiDKDb0bGjoHb3+SU/VvVowMrNjLYMVXwidBAXaiEuxEJXwSFPCJl4MbL0XOqRR0K/72zHFl6/cPDZtnFgx+CruWu7VmP1epjvD7eRAURVEbI4p/tylKmsaIknUyIqU/sGJkeDUZkdIfGDHSa97RUtGGfSW/f70+h6LWqw5wFOoIP8jDfOYqeCyvNUMsRVDOei++ciMrQR3A4tNbWQm0FxWUs361shyKWl8ZzlGWhvqA3s8O//kAvyBoHu9NOpzlC4p6438C8Hr8CN553KkxVTnMAAAAAElFTkSuQmCC';
 const base64MultiSearchIcon =
@@ -33,6 +34,7 @@ let textSelection = '';
 let navEntered = false;
 let xPos;
 let yPos;
+let targetUrlSent = false;
 
 // Current state
 if (logToConsole) {
@@ -205,7 +207,7 @@ async function init() {
         }
     }
 
-
+    // If debugging mode is enabled, log the tab url and domain
     if (logToConsole) {
         console.log(`Tab url: ${tabUrl}`);
         console.log(`Path name: ${pn}`);
@@ -513,23 +515,80 @@ function getSelectionEndPosition() {
     return { x: 0, y: 0 };
 }
 
-function handleRightClickWithoutGrid(e) {
-    if (logToConsole) console.log(e);
-    const selectedText = getSelectedText();
-    if (logToConsole) console.log(selectedText);
-    // Send the selected text to background.js
-    sendMessage('setSelection', { selection: selectedText });
-    // If right click is on image
-    const elementClicked = e.target;
-    const tag = elementClicked.tagName;
-    if (tag === 'IMG') {
-        if (window.getSelection) {
-            window.getSelection().removeAllRanges();
+async function handleRightClickWithoutGrid(e) {
+    if (logToConsole) console.log(`Target url sent: ${targetUrlSent}`);
+
+    // If the target url has already been sent then do nothing
+    if (!targetUrlSent) {
+        //e.preventDefault();
+        const { clientX: x, clientY: y } = e;
+        if (logToConsole) console.log(e);
+
+        // If right click is on image
+        const elementClicked = e.target;
+        const tag = elementClicked.tagName;
+        const className = elementClicked.className;
+        if (tag === 'IMG' || (tag === 'DIV' && className.includes('div.iris-annotation-layer'))) {
+            if (domain.includes('youtube.com') || domain.includes('youtu.be') || domain.includes('youtube-nocookie.com') || domain.includes('vimeo.com')) {
+                // Get the video url
+                const videoUrl = absoluteUrl(getClosestAnchorHref(elementClicked));
+                //const videoId = new URL(videoUrl).searchParams.get('v');
+                //const downloadUrl = ytDownloadUrl + videoId;
+                await sendMessage('setTargetUrl', videoUrl);
+                if (logToConsole) console.log(`Video url: ${videoUrl}`);
+            } else {
+                if (window.getSelection) {
+                    window.getSelection().removeAllRanges();
+                }
+                // Get the image url
+                const imgUrl = absoluteUrl(elementClicked.getAttribute('src'));
+                await sendMessage('setTargetUrl', imgUrl);
+                if (logToConsole) console.log(`Image url: ${imgUrl}`);
+            }
+            targetUrlSent = true;
+        } else {
+            const selectedText = getSelectedText();
+            if (logToConsole) console.log(selectedText);
+            // Send the selected text to background.js
+            await sendMessage('setSelection', { selection: selectedText });
         }
-        const imgUrl = absoluteUrl(elementClicked.getAttribute('src'));
-        sendMessage('setTargetUrl', imgUrl);
-        if (logToConsole) console.log(`Image url: ${imgUrl}`);
+
+        // Dispatch the new event on the original target element
+        if (targetUrlSent) {
+
+            // Dispatch the new event on the original target element
+            setTimeout(() => {
+                // Create a new context menu event
+                const newEvent = new MouseEvent('contextmenu', {
+                    bubbles: true,
+                    cancelable: false,
+                    view: window,
+                    button: 2,
+                    buttons: 2,
+                    clientX: x,
+                    clientY: y
+                });
+
+                // Dispatch the new event
+                elementClicked.dispatchEvent(newEvent);
+                if (logToConsole) console.log('New contextmenu event fired.');
+                if (logToConsole) console.log(`Default prevented: ${newEvent.defaultPrevented}`);
+                if (logToConsole) console.log(newEvent.target);
+            }, 100); // Small delay to ensure proper event handling
+        }
+
+    } else {
+        targetUrlSent = false;
     }
+}
+
+function getClosestAnchorHref(imgElement) {
+    if (!imgElement || imgElement.tagName !== 'IMG') {
+        throw new Error('Provided element is not an img tag');
+    }
+
+    const anchorElement = imgElement.closest('a');
+    return anchorElement ? anchorElement.href : null;
 }
 
 // Display clickable buttons/icons on mycroftproject.com